Add files via upload
The NIM game is done! :D
after some work (and a lot of time) I finally did this game/exercise from study grid.

this program takes 2 numbers input, one is the number of pieces in the board, and the other is the limit that one player can remove in his turn.